requirements.txt
langchain==0.3.27
langchain-community==0.3.30
jinja2==3.1.6
requests==2.32.5
python-dotenv==1.1.1
openai==1.100.2
bs4==4.12.3
selenium==4.24.0